refactor(tenant): enhance tenant management and deposit handling
Key Changes:
- Renamed moveOutTenant.ts to endTenantLease.ts for better clarity
- Updated recordDeposit to return deposit data in a structured format
- Enhanced error handling and response formatting across tenant controllers
- Improved validation schemas for tenant-related operations
- Updated database schema with new migrations for deposit tracking
- Refactored tenant search and filtering functionality
- Added new endpoints for deposit management (record, refund, list)
- Improved property search with additional filters and sorting
- Updated API routes to reflect new tenant management structure
- Enhanced security by ensuring proper ownership verification
- Added support for tenant document uploads
- Improved family member management for tenants
- Updated maintenance history tracking
- Enhanced payment history and reporting
- Added support for tenant favorites and property management
- Improved error handling and validation across all endpoints

Database Changes:
- Added new tables/indexes for deposit tracking
- Updated tenant schema with new fields
- Added support for document storage
- Improved data relationships and constraints

Security:
- Added ownership verification for all tenant operations
- Improved input validation
- Enhanced error messages to avoid information leakage
- Added proper access controls for all endpoints

This commit significantly improves the tenant management system with better error handling, security, and functionality for deposit management and lease operations.

+++ src/controllers/api/media/uploadMedia.ts
@@ -52,7 +52,7 @@ const uploadMedia: RequestHandler = async (request, response) => {
 
     // Check existing media count
     const [existingMediaCount] = await db
-      .select({ count: sql<number>`count(*)` })
+      .select({ count: sql<number>`count(*)::int` })
       .from(media)
       .where(eq(media.propertyId, propertyId));
